// src/types/api/service/models/ModifyNoticePolicyDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ModifyNoticePolicyDto
 */
export interface ModifyNoticePolicyDto {
    /**
     * 이메일 알림 활성화 여부
     * @type {boolean}
     * @memberof ModifyNoticePolicyDto
     */
    flagActive: boolean;
    /**
     * 내가 댓글을 단 콘텐츠에 누군가가 댓글을 달았을 때
     * @type {boolean}
     * @memberof ModifyNoticePolicyDto
     */
    addCommentAddCommentContent: boolean;
    /**
     * 내가 댓글을 단 콘텐츠를 누군가 수정했을 때
     * @type {boolean}
     * @memberof ModifyNoticePolicyDto
     */
    modifyContentAddCommentContent: boolean;
    /**
     * 내가 생성하거나 수정한 콘텐츠를 누군가 수정했을 때
     * @type {boolean}
     * @memberof ModifyNoticePolicyDto
     */
    modifyContentCuContent: boolean;
    /**
     * 내가 소유한 콘텐츠에 누군가 댓글을 달았을 때
     * @type {boolean}
     * @memberof ModifyNoticePolicyDto
     */
    addCommentOwnerContent: boolean;
    /**
     * 내가 소유한 콘텐츠를 누군가 수정했을 때
     * @type {boolean}
     * @memberof ModifyNoticePolicyDto
     */
    modifyContentOwnerContent: boolean;
    /**
     * 내 회원 정보나 권한, 그룹이 변경되었을때
     * @type {boolean}
     * @memberof ModifyNoticePolicyDto
     */
    changeUserInfo: boolean;
}

/**
 * Check if a given object implements the ModifyNoticePolicyDto interface.
 */
export function instanceOfModifyNoticePolicyDto(value: object): value is ModifyNoticePolicyDto {
    if (!('flagActive' in value) || value['flagActive'] === undefined) return false;
    if (!('addCommentAddCommentContent' in value) || value['addCommentAddCommentContent'] === undefined) return false;
    if (!('modifyContentAddCommentContent' in value) || value['modifyContentAddCommentContent'] === undefined) return false;
    if (!('modifyContentCuContent' in value) || value['modifyContentCuContent'] === undefined) return false;
    if (!('addCommentOwnerContent' in value) || value['addCommentOwnerContent'] === undefined) return false;
    if (!('modifyContentOwnerContent' in value) || value['modifyContentOwnerContent'] === undefined) return false;
    if (!('changeUserInfo' in value) || value['changeUserInfo'] === undefined) return false;
    return true;
}

export function ModifyNoticePolicyDtoFromJSON(json: any): ModifyNoticePolicyDto {
    return ModifyNoticePolicyDtoFromJSONTyped(json, false);
}

export function ModifyNoticePolicyDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ModifyNoticePolicyDto {
    if (json == null) {
        return json;
    }
    return {
        
        'flagActive': json['flagActive'],
        'addCommentAddCommentContent': json['addCommentAddCommentContent'],
        'modifyContentAddCommentContent': json['modifyContentAddCommentContent'],
        'modifyContentCuContent': json['modifyContentCuContent'],
        'addCommentOwnerContent': json['addCommentOwnerContent'],
        'modifyContentOwnerContent': json['modifyContentOwnerContent'],
        'changeUserInfo': json['changeUserInfo'],
    };
}

export function ModifyNoticePolicyDtoToJSON(value?: ModifyNoticePolicyDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'flagActive': value['flagActive'],
        'addCommentAddCommentContent': value['addCommentAddCommentContent'],
        'modifyContentAddCommentContent': value['modifyContentAddCommentContent'],
        'modifyContentCuContent': value['modifyContentCuContent'],
        'addCommentOwnerContent': value['addCommentOwnerContent'],
        'modifyContentOwnerContent': value['modifyContentOwnerContent'],
        'changeUserInfo': value['changeUserInfo'],
    };
}