// src/types/api/service/models/NoticePolicyDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface NoticePolicyDto
 */
export interface NoticePolicyDto {
    /**
     * 
     * @type {number}
     * @memberof NoticePolicyDto
     */
    noticePolicyId?: number;
    /**
     * 알림 활성화 여부
     * @type {boolean}
     * @memberof NoticePolicyDto
     */
    flagActive?: boolean;
    /**
     * 내가 댓글을 단 콘텐츠에 누군가가 댓글을 달았을 때
     * @type {boolean}
     * @memberof NoticePolicyDto
     */
    addCommentAddCommentContent?: boolean;
    /**
     * 내가 댓글을 단 콘텐츠를 누군가 수정했을 때
     * @type {boolean}
     * @memberof NoticePolicyDto
     */
    modifyContentAddCommentContent?: boolean;
    /**
     * 내가 생성하거나 수정한 콘텐츠를 누군가 수정했을 때
     * @type {boolean}
     * @memberof NoticePolicyDto
     */
    modifyContentCuContent?: boolean;
    /**
     * 내가 소유한 콘텐츠에 누군가 댓글을 달았을 때
     * @type {boolean}
     * @memberof NoticePolicyDto
     */
    addCommentOwnerContent?: boolean;
    /**
     * 내가 소유한 콘텐츠를 누군가 수정했을 때
     * @type {boolean}
     * @memberof NoticePolicyDto
     */
    modifyContentOwnerContent?: boolean;
    /**
     * 내 회원 정보나 권한, 그룹이 변경되었을때
     * @type {boolean}
     * @memberof NoticePolicyDto
     */
    changeUserInfo?: boolean;
    /**
     * 사용자 Id
     * @type {number}
     * @memberof NoticePolicyDto
     */
    userId?: number;
}

/**
 * Check if a given object implements the NoticePolicyDto interface.
 */
export function instanceOfNoticePolicyDto(value: object): value is NoticePolicyDto {
    return true;
}

export function NoticePolicyDtoFromJSON(json: any): NoticePolicyDto {
    return NoticePolicyDtoFromJSONTyped(json, false);
}

export function NoticePolicyDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): NoticePolicyDto {
    if (json == null) {
        return json;
    }
    return {
        
        'noticePolicyId': json['noticePolicyId'] == null ? undefined : json['noticePolicyId'],
        'flagActive': json['flagActive'] == null ? undefined : json['flagActive'],
        'addCommentAddCommentContent': json['addCommentAddCommentContent'] == null ? undefined : json['addCommentAddCommentContent'],
        'modifyContentAddCommentContent': json['modifyContentAddCommentContent'] == null ? undefined : json['modifyContentAddCommentContent'],
        'modifyContentCuContent': json['modifyContentCuContent'] == null ? undefined : json['modifyContentCuContent'],
        'addCommentOwnerContent': json['addCommentOwnerContent'] == null ? undefined : json['addCommentOwnerContent'],
        'modifyContentOwnerContent': json['modifyContentOwnerContent'] == null ? undefined : json['modifyContentOwnerContent'],
        'changeUserInfo': json['changeUserInfo'] == null ? undefined : json['changeUserInfo'],
        'userId': json['userId'] == null ? undefined : json['userId'],
    };
}

export function NoticePolicyDtoToJSON(value?: NoticePolicyDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'noticePolicyId': value['noticePolicyId'],
        'flagActive': value['flagActive'],
        'addCommentAddCommentContent': value['addCommentAddCommentContent'],
        'modifyContentAddCommentContent': value['modifyContentAddCommentContent'],
        'modifyContentCuContent': value['modifyContentCuContent'],
        'addCommentOwnerContent': value['addCommentOwnerContent'],
        'modifyContentOwnerContent': value['modifyContentOwnerContent'],
        'changeUserInfo': value['changeUserInfo'],
        'userId': value['userId'],
    };
}